requirements.txt
Django==2.1.7
django-cors-headers==2.5.2
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
dnspython==1.16.0
mongoengine==0.16.3
pymongo==3.7.2

